refactor: pass decode jwt in verify page as props for json block and credential tabs
