+++ packages/mock-app/src/components/CredentialRender/CredentialRender.tsx
@@ -1,13 +1,13 @@
 import React, { useCallback, useEffect, useState } from 'react';
 import { Renderer, WebRenderingTemplate2022 } from '@vckit/renderer';
-import { VerifiableCredential } from '@vckit/core-types';
+import { UnsignedCredential, VerifiableCredential } from '@vckit/core-types';
 import { Box, CircularProgress } from '@mui/material';
 import { convertBase64ToString } from '../../utils';
 
 /**
  * CredentialRender component is used to render the credential
  */
-const CredentialRender = ({ credential }: { credential: VerifiableCredential }) => {
+const CredentialRender = ({ credential }: { credential: VerifiableCredential | UnsignedCredential }) => {
   const [documents, setDocuments] = useState<string[]>([]);
   const [isLoading, setIsLoading] = useState<boolean>(false);
