+++ packages/mock-app/src/components/CredentialTabs/CredentialTabs.tsx
@@ -1,35 +1,33 @@
-import React, { useContext, useEffect } from 'react';
-import { Box, Button, Tab, Tabs, useMediaQuery, useTheme } from '@mui/material';
-import { VerifiableCredential } from '@vckit/core-types';
-import CloudDownloadOutlinedIcon from '@mui/icons-material/CloudDownloadOutlined';
+import React, { useEffect } from 'react';
+import { Box, Tab, Tabs, useMediaQuery, useTheme } from '@mui/material';
 
 import CredentialRender from '../CredentialRender/CredentialRender';
 import { JsonBlock } from '../JsonBlock';
-import { VerifyPageContext } from '../../hooks/VerifyPageContext';
+import { CredentialComponentProps } from '../../types/common.types';
+import { DownloadCredentialButton } from '../DownloadCredentialButton/DownloadCredentialButton';
 
-const CredentialTabs = ({ credential }: { credential: VerifiableCredential }) => {
+const CredentialTabs = ({ credential, decodeCredential }: CredentialComponentProps) => {
   const credentialTabs = [
     {
       label: 'Rendered',
-      children: <CredentialRender credential={credential} />,
+      children: <CredentialRender credential={decodeCredential ?? credential} />,
     },
     {
       label: 'JSON',
-      children: <JsonBlock credential={credential} />,
+      children: <JsonBlock credential={decodeCredential ?? credential} />,
     },
   ];
 
   const [currentTabIndex, setCurrentTabIndex] = React.useState(0);
   const theme = useTheme();
   const isMobile = useMediaQuery(theme.breakpoints.down('sm'));
-  const { vc } = useContext(VerifyPageContext);
 
   useEffect(() => {
     configDefaultTabs();
   }, [credential]);
 
   const configDefaultTabs = () => {
-    if (credential?.render?.[0]?.template) {
+    if (decodeCredential?.render?.[0]?.template) {
       return setCurrentTabIndex(0);
     }
 
@@ -39,20 +37,6 @@ const CredentialTabs = ({ credential }: { credential: VerifiableCredential }) =>
   const handleChange = (event: React.SyntheticEvent, newValue: number) => {
     event.preventDefault();
     setCurrentTabIndex(newValue);
-  };
-
-  /**
-   * handle click on download button
-   */
-  const handleClickDownloadVC = async () => {
-    const element = document.createElement('a');
-    const file = new Blob([JSON.stringify({ verifiableCredential: vc }, null, 2)], {
-      type: 'text/plain',
-    });
-    element.href = URL.createObjectURL(file);
-    element.download = 'vc.json';
-    document.body.appendChild(element); // Required for this to work in FireFox
-    element.click();
   };
 
   const TabPanel = ({ children, value, index, ...other }: any) => (
@@ -93,22 +77,7 @@ const CredentialTabs = ({ credential }: { credential: VerifiableCredential }) =>
         </Tabs>
 
         {/* Download Button */}
-        <Button
-          variant='text'
-          startIcon={<CloudDownloadOutlinedIcon sx={{ marginRight: '5px' }} />}
-          sx={{
-            color: 'primary.main',
-            textTransform: 'none',
-            marginLeft: 2,
-            paddingRight: 0,
-            justifyContent: 'end',
-            fontSize: '16px',
-            '.MuiButton-startIcon': { marginRight: 0 },
-          }}
-          onClick={handleClickDownloadVC}
-        >
-          {isMobile ? '' : 'Download'}
-        </Button>
+        <DownloadCredentialButton credential={credential} />
       </Box>
 
       {/* Tab Panels */}
