+++ packages/mock-app/src/components/DownloadCredentialButton/index.ts
@@ -1,0 +1,1 @@
+export * from './DownloadCredentialButton';
