+++ packages/mock-app/src/pages/Verify.tsx
@@ -11,7 +11,6 @@ import { LoadingWithText } from '../components/LoadingWithText';
 import { BackButton } from '../components/BackButton';
 import Credential from '../components/Credential/Credential';
 import appConfig from '../constants/app-config.json';
-import { VerifyPageContext } from '../hooks/VerifyPageContext';
 
 enum PassportStatus {
   'LOADING_FETCHING_PASSPORT' = 'LOADING_FETCHING_PASSPORT',
@@ -160,11 +159,9 @@ const Verify = () => {
         }
 
         return (
-          <VerifyPageContext.Provider value={{ vc: credential }}>
-            <BackButton>
-              <Credential credential={customCredential ?? credential} />
-            </BackButton>
-          </VerifyPageContext.Provider>
+          <BackButton>
+            <Credential credential={credential} decodeCredential={customCredential} />
+          </BackButton>
         );
       default:
         return (
